+++ just_audio/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.9.30
 
 * Upgrade ExoPlayer to 2.18.1.
+* Fix bug using headers with LockCachingAudioSource.
 
 ## 0.9.29